bybit_bot/requirements.txt
# ═══════════════════════════════════════
# Bybit Bot - PRE-COMPILED VERSION
# بدون نیاز به Visual Studio
# ═══════════════════════════════════════

# Core Framework
streamlit==1.28.0

# Data & Charts (نسخه‌های قبلی با wheel آماده)
pandas==2.0.3
numpy==1.24.4
plotly==5.17.0

# API
requests==2.31.0
urllib3==2.0.7

# DateTime
python-dateutil==2.8.2
pytz==2023.3

# WebSocket
websocket-client==1.6.4

# Excel (optional)
openpyxl==3.1.2

# ═══════════════════════════════════════
# این نسخه‌ها pre-compiled هستن
# نیازی به Visual Studio ندارن
# ═══════════════════════════════════════